+++ programs/whirlpool/src/util/mod.rs
@@ -3,6 +3,7 @@ pub mod sparse_swap;
 pub mod swap_tick_sequence;
 pub mod swap_utils;
 pub mod token;
+pub mod token_2022;
 pub mod v2;
 
 pub use shared::*;
@@ -10,6 +11,7 @@ pub use sparse_swap::*;
 pub use swap_tick_sequence::*;
 pub use swap_utils::*;
 pub use token::*;
+pub use token_2022::*;
 pub use v2::*;
 
 #[cfg(test)]
